+++ lib/kitchen/patches/integer.rb
@@ -3,7 +3,10 @@
 # Monkey patches for +Integer+
 #
 class Integer
-  ROMAN_NUMERALS = %w[0 i ii iii iv v vi vii viii ix x xi xii xiii xiv xv xvi xvii xviii xix xx].freeze
+  ROMAN_NUMERALS = %w[0 i ii iii iv v vi vii viii ix
+                      x xi xii xiii xiv xv xvi xvii xviii xix
+                      xx xxi xxii xxiii xxiv xxv xxvi xxvii xxviii xxix
+                      xxx xxxi xxxii xxxiii xxxiv xxxv xxxvi xxxvii xxxviii xxxix].freeze
 
   # Formats as different types of integers, including roman numerals.
   #